+++ CHANGES.rst
@@ -31,6 +31,9 @@ Changelog
   [chris-adam, sgeulette]
 - Added possibility to have elements of the same context to belong to different sessions.
   [chris-adam]
+- Configured the `@@remove-from-esign-session` the same way the `@@remove-item-from-esign-session`
+  so relying on an `available` method to show it only if a context in a session.
+  [gbastien]
 
 1.0a2 (2026-02-06)
 ------------------